# [fips-legacy-8] perf: Disallow mis-matched inherited group reads